refactor: directory structure for components
