+++ src/App.css
@@ -1,8 +1,8 @@
 @import url('components/Page/index.css');
 @import url('components/Page/Navbar/index.css');
 @import url('pages/ListPage/index.css');
-@import url('pages/ListPage/CategoryNavbar.css');
-@import url('pages/ListPage/CategoryNavbarTab.css');
+@import url('components/ListPage/CategoryNavbar/index.css');
+@import url('components/ListPage/CategoryNavbarTab/index.css');
 
 body {
   margin: 0;
