+++ src/core/router/index.ts
@@ -3,9 +3,9 @@ import { Router } from './types/router';
 /**
  * 라우터 생성에 필요한 속성을 정의합니다.
  * @interface CreateRouterProps
- * @property {{ [key: string]: HTMLElement }} routes - 경로와 해당 컴포넌트 매핑
+ * @property {{ [key: string]: () => HTMLElement }} routes - 경로와 해당 컴포넌트 매핑
  * @property {HTMLElement} root - 라우터가 작동할 루트 엘리먼트
- * @property {HTMLElement} errorPage - 에러 페이지 엘리먼트
+ * @property {() => HTMLElement} errorPage - 에러 페이지를 반환하는 함수
  */
 interface CreateRouterProps {
   routes: { [key: string]: () => HTMLElement };
